// backend/src/common/enums/action-type.enum.ts
export enum ActionType {
  CREATE = 'CREATE',
  SUBMIT = 'SUBMIT',
  ASSIGN = 'ASSIGN',
  RETURN = 'RETURN',
  REVIEW = 'REVIEW',
  CLOSE = 'CLOSE',
  CORRECT_MATERIALS = 'CORRECT_MATERIALS',
  CORRECT_EVIDENCE = 'CORRECT_EVIDENCE',
  CORRECT_INFO = 'CORRECT_INFO',
  ADD_ATTACHMENT = 'ADD_ATTACHMENT',
  ADD_AUDIT_NOTE = 'ADD_AUDIT_NOTE',
}